+++ routes/leaderRouter.js
@@ -19,27 +19,39 @@ leaderRouter.route('/')
                   }, (err) => next(err))
                   .catch((err) => next(err));
       })
-      .post(authenticate.verifyUser, (req, res, next) => {
-            Leaders.create(req.body)
-                  .then((leader) => {
-                        res.statusCode = 200;
-                        res.setHeader('Content-type', 'application/json');
-                        res.json(leader);
-                  }, (err) => next(err))
-                  .catch((err) => next(err));
+      .post(authenticate.verifyAdmin, (req, res, next) => {
+            if (req.user.admin) {
+                  Leaders.create(req.body)
+                        .then((leader) => {
+                              res.statusCode = 200;
+                              res.setHeader('Content-type', 'application/json');
+                              res.json(leader);
+                        }, (err) => next(err))
+                        .catch((err) => next(err));
+            } else {
+                  err = new Error('Not Authorized :: not admin')
+                  err.status = 404;
+                  return next(err)
+            }
       })
       .put(authenticate.verifyUser, (req, res, next) => {
             res.statusCode = 403;
             res.end('PUT operation not supported on /leaders');
       })
-      .delete(authenticate.verifyUser, (req, res, next) => {
-            Leaders.remove({})
-                  .then((resp) => {
-                        res.statusCode = 200;
-                        res.setHeader('Content-type', 'application/json');
-                        res.json(resp);
-                  }, (err) => next(err))
-                  .catch((err) => next(err));
+      .delete(authenticate.verifyAdmin, (req, res, next) => {
+            if (req.user.admin) {
+                  Leaders.remove({})
+                        .then((resp) => {
+                              res.statusCode = 200;
+                              res.setHeader('Content-type', 'application/json');
+                              res.json(resp);
+                        }, (err) => next(err))
+                        .catch((err) => next(err));
+            } else {
+                  err = new Error('Not Authorized :: not admin')
+                  err.status = 404;
+                  return next(err)
+            }
       });
 
 /***************LEADERS ID***************/
@@ -57,27 +69,39 @@ leaderRouter.route('/:leaderId')
             res.statusCode = 403;
             res.end('POST operation not supported on /promotions');
       })
-      .put(authenticate.verifyUser, (req, res, next) => {
-            Leaders.findByIdAndUpdate(req.params.leaderId, {
-                        $set: req.body
-                  }, {
-                        new: true
-                  })
-                  .then((resp) => {
-                        res.StatusCode = 200;
-                        res.setHeader('Content-type', 'application/json');
-                        res.json(resp);
-                  }, (err) => next(err))
-                  .catch((err) => next(err));
+      .put(authenticate.verifyAdmin, (req, res, next) => {
+            if (req.user.admin) {
+                  Leaders.findByIdAndUpdate(req.params.leaderId, {
+                              $set: req.body
+                        }, {
+                              new: true
+                        })
+                        .then((resp) => {
+                              res.StatusCode = 200;
+                              res.setHeader('Content-type', 'application/json');
+                              res.json(resp);
+                        }, (err) => next(err))
+                        .catch((err) => next(err));
+            } else {
+                  err = new Error('Not Authorized :: not admin')
+                  err.status = 404;
+                  return next(err)
+            }
       })
-      .delete(authenticate.verifyUser, (req, res, next) => {
-            Leaders.findByIdAndRemove(req.params.leaderId)
-                  .then((resp) => {
-                        res.StatusCode = 200;
-                        res.setHeader('Content-type', 'application/json');
-                        res.json(resp);
-                  }, (err) => next(err))
-                  .catch((err) => next(err));
+      .delete(authenticate.verifyAdmin, (req, res, next) => {
+            if (req.user.admin) {
+                  Leaders.findByIdAndRemove(req.params.leaderId)
+                        .then((resp) => {
+                              res.StatusCode = 200;
+                              res.setHeader('Content-type', 'application/json');
+                              res.json(resp);
+                        }, (err) => next(err))
+                        .catch((err) => next(err));
+            } else {
+                  err = new Error('Not Authorized :: not admin')
+                  err.status = 404;
+                  return next(err)
+            }
       });
 
 module.exports = leaderRouter;
